move in new package /pkg/nix
